Namespacing tables to have gutentag_ prefix.
This could have been a migration, but then people installing Gutentag for the first time who already have tables named 'tags' or 'taggings' would have migrations complain. So, renaming everything in existing migrations instead (for people starting now with Gutentag), and there'll be a note in the README for anyone upgrading.