Pull logic of adding validations from tag model.
Related to #17, this is so if others modify the behaviour of the tag model, they can change the validations that are in play. It's worth noting that adding a custom tag_validations class/object/proc (whatever it is, it must respond to call and take the tag model as an argument) will only work in Rails, due to the lazy loading of models. In non-Rails apps, the model will be loaded with the rest of the gem, and changing tag_validations will have no impact. I'm open to suggestions on how to better manage this.