Separating persistence from changed state.
Influenced by the suggestion of @pboling in #39, this commit introduces Gutentag::ChangeState, which contains the information about added and removed tags from a given taggable instance. It is then used by Gutentag::Persistence to save said changes to the database.

If you wanted to add some logic around the persistence, my current recommended approach would be to overwrite the persist_tags method in your taggable model, create the Gutentag::ChangeState instance, pass it to Gutentag::Persistence, and then also do other things with it.