Fix `tag_names` attribute in Rails 4.2.
As raised in #67 - `tag_names` was being treated as a database column, which is not correct. A regression spec has been added, and the handling of dirty state for the attribute has been reworked accordingly.

It’s not particularly clean, but it works, and I guess this is the price of maintaining support across so many versions of Rails.