feat: Tweak migrations to current Rails version.
As discussed in #80, the current approach to migrations in engines is flawed. So, instead of the clunky check in each migration, we can instead use a generator to update them to use the current Rails version at the time of the migrations being added. The test suite uses this approach too, so we can be sure that the migrations work in each Rails release. Essentially, this automates the recommendation in 04334272e41c55fb679c6aa03c5fc8eac0ec04cb.

It's not quite perfect - if new migrations get added to Gutentag, and someone's upgrading their app, the generator will be invoked on all Gutentag migrations (so, editing the old ones to use a potentially newer version of Rails). But at this point in time I'm not expecting further migrations - so, we can look into improving this if that changes.