feat: have migrations versioned by default.
Instead of defaulting to no AR version in the migration files, let's
instead use 4.2 as the default version - so they work out of the box in
modern (v5+) Rails applications without the need for the generator.

The generator, if run, will then update to the current Rails version
(which includes removing the version if on a pre-modern release).

The generator is still recommended, as the migrations are confirmed to
work on known releases, and it ensures the defaults (such as 64-bit
integer primary keys) are used from the get-go.